move timestamp parse
